requirements/base.txt
Django==1.7
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
Pygments==1.6
South==0.8.4
beautifulsoup4==4.3.2
django-filter==0.7
django-wpadmin==1.7.0
djangorestframework==2.4.2
gunicorn==19.1.1
python-dateutil==2.3
six==1.8.0
urwid==1.3.0
wsgiref==0.1.2
